+++ lib/src/dart_formatter.dart
@@ -206,8 +206,10 @@ class DartFormatter {
     var featureSet = FeatureSet.fromEnableFlags2(
       sdkLanguageVersion: Version(2, 19, 0),
       flags: [
+        'class-modifiers',
         if (patterns) 'patterns',
         'records',
+        'sealed-class',
         'unnamed-libraries',
       ],
     );