+++ lib/src/debug.dart
@@ -126,25 +126,30 @@ void dumpChunks(int start, List<Chunk> chunks) {
       var spanBars = '';
       for (var span in spans) {
         if (chunk.spans.contains(span)) {
-          if (index == 0 || !chunks[index - 1].spans.contains(span)) {
+          if (index == chunks.length - 1 ||
+              !chunks[index + 1].spans.contains(span)) {
+            // This is the last chunk with the span.
+            spanBars += '╙';
+          } else {
+            spanBars += '║';
+          }
+        } else {
+          // If the next chunk has this span, then show it bridging this chunk
+          // and the next because a split between them breaks the span.
+          if (index < chunks.length - 1 &&
+              chunks[index + 1].spans.contains(span)) {
             if (span.cost == 1) {
               spanBars += '╓';
             } else {
               spanBars += span.cost.toString();
             }
-          } else {
-            spanBars += '║';
-          }
-        } else {
-          if (index > 0 && chunks[index - 1].spans.contains(span)) {
-            spanBars += '╙';
-          } else {
-            spanBars += ' ';
           }
         }
       }
       row.add(spanBars);
     }
+
+    row.add(chunk.spans.map((span) => span.id).join(' '));
 
     if (chunk.text.length > 70) {
       row.add(chunk.text.substring(0, 70));