+++ lib/src/source_visitor.dart
@@ -583,6 +583,11 @@ class SourceVisitor extends ThrowingAstVisitor {
 
     builder.nestExpression();
     modifier(node.abstractKeyword);
+    modifier(node.baseKeyword);
+    modifier(node.interfaceKeyword);
+    modifier(node.finalKeyword);
+    modifier(node.sealedKeyword);
+    modifier(node.mixinKeyword);
     token(node.classKeyword);
     space();
     token(node.name);
@@ -602,6 +607,11 @@ class SourceVisitor extends ThrowingAstVisitor {
 
     _simpleStatement(node, () {
       modifier(node.abstractKeyword);
+      modifier(node.baseKeyword);
+      modifier(node.interfaceKeyword);
+      modifier(node.finalKeyword);
+      modifier(node.sealedKeyword);
+      modifier(node.mixinKeyword);
       token(node.typedefKeyword);
       space();
       token(node.name);
@@ -2203,6 +2213,10 @@ class SourceVisitor extends ThrowingAstVisitor {
     visitMetadata(node.metadata);
 
     builder.nestExpression();
+    modifier(node.baseKeyword);
+    modifier(node.interfaceKeyword);
+    modifier(node.finalKeyword);
+    modifier(node.sealedKeyword);
     token(node.mixinKeyword);
     space();
     token(node.name);